@@ -8,13 +8,13 @@ Bundler.require(*Rails.groups)
 
 module AntCat
   class Application < Rails::Application
-    # Set Time.zone default to the specified zone and make Active Record auto-convert to this zone.
-    # Run "rake -D time" for a list of tasks for finding time zone names. Default is UTC.
-    # config.time_zone = 'Central Time (US & Canada)'
+    # Initialize configuration defaults for originally generated Rails version.
+    #config.load_defaults 5.1
 
-    # The default locale is :en and all translations from config/locales/*.rb,yml are auto loaded.
-    # config.i18n.load_path += Dir[Rails.root.join('my', 'locales', '*.{rb,yml}').to_s]
-    # config.i18n.default_locale = :de
+    # Settings in config/environments/* take precedence over those specified here.
+    # Application configuration should go into files in config/initializers
+    # -- all .rb files in that directory are automatically loaded.
+
     config.action_mailer.delivery_method = :sendmail
     config.action_mailer.sendmail_settings = { arguments: '-i' }
     config.active_record.observers = [
@@ -26,6 +26,8 @@ module AntCat
       :reference_document_observer,
       :reference_observer,
     ]
+
+    config.enable_dependency_loading = true
 
     config.autoload_paths += Dir["#{config.root}/lib/**/"]
     config.autoload_paths += Dir["#{config.root}/lib/tasks/**"]
